try consolidating more methods
